Split measure functions out of the measure example
